backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.23
pydantic>=2.12.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
openai>=1.3.5
psycopg2-binary
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.2
